+++ lib/controllers/fav_caption_controller.dart
@@ -1,10 +1,14 @@
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:unicap_cg/common/basewidgets/custom_toast_message.dart';
 import 'package:unicap_cg/data/local/databse_helper.dart';
 import 'package:unicap_cg/models/caption_category.dart';
 import 'package:unicap_cg/models/fav_caption_model.dart';
 import 'package:unicap_cg/services/fav_caption_service.dart';
 
 class FavoriteCaptionController with ChangeNotifier {
+  final Connectivity _connectivity = Connectivity();
   final FavoriteCaptionService _service = FavoriteCaptionService();
   final DatabaseHelper _dbHelper = DatabaseHelper();
   List<FavoriteCaption> _favorites = [];
@@ -12,7 +16,9 @@ class FavoriteCaptionController with ChangeNotifier {
   List<FavoriteCaption>? get favorites => _favorites;
 
   Future<void> syncUnsyncedFavCaptions(String userId) async {
-    await _service.syncFavoriteCaptions(userId);
+    if(await isUserOnline()){
+      await _service.syncFavoriteCaptions(userId);
+    }
     await loadAndSaveFavCaptions(userId);
   }
 
@@ -21,7 +27,11 @@ class FavoriteCaptionController with ChangeNotifier {
     // notifyListeners();
 
     try {
-      List<FavoriteCaption>? allFavCaptions = await _service.getFavCaptions(userId);
+      List<FavoriteCaption>? allFavCaptions;
+
+      if(await isUserOnline()){
+        allFavCaptions = await _service.getFavCaptions(userId);
+      }
 
 
       if(allFavCaptions?.isNotEmpty ?? false){
@@ -31,7 +41,7 @@ class FavoriteCaptionController with ChangeNotifier {
       }
 
       _favorites = await _dbHelper.getFavoriteCaptions(userId);
-
+      notifyListeners();
       print('=====favorites=========${favorites?.length}==============');
 
     } catch (e) {
@@ -44,7 +54,7 @@ class FavoriteCaptionController with ChangeNotifier {
   }
 
   Future<void> addFavorite(String userId, Caption caption) async {
-    print('============Inside addFavorit=========');
+    print('============Inside addFavorite=========');
     try {
       print('============Inside try=========');
       await _dbHelper.insertFavoriteCaption(FavoriteCaption(
@@ -54,40 +64,52 @@ class FavoriteCaptionController with ChangeNotifier {
         isSynced: false,
       ));
 
+    } catch (e) {
+      print('==============Error Inserting favorite favorite to local: ================$e');
+    }
+
+    if(await isUserOnline()){
       await _service.addFavoriteCaption(FavoriteCaption(
         userId: userId,
         captionId: caption.key,
         caption: caption.caption,
         isSynced: false,
       ));
-
-      await getAllFavCaptionsFromLocal(userId);
-
-    } catch (e) {
-      print('==============Error Inserting favorite favorite: ================$e');
-      rethrow;
     }
+
+
+    await getAllFavCaptionsFromLocal(userId);
+  }
+
+  Future<bool> isUserOnline() async {
+    final connectivityResult = await _connectivity.checkConnectivity();
+    final bool isUserOnline = connectivityResult != ConnectivityResult.none;
+    return isUserOnline;
   }
 
   Future<void> removeFavorite(String userId, String captionId) async {
     try {
-      await _service.removeFavoriteCaption(userId, captionId);
+      print('============inside remove try==============$userId and $captionId}');
       await _dbHelper.deleteFavoriteCaption(userId, captionId);
-      await loadAndSaveFavCaptions(userId);
-      notifyListeners();
     } catch (e) {
-      print('Error removing favorite: $e');
-      rethrow;
+      print('=============controller============Error removing favorite from local: $e');
     }
+
+    if(await isUserOnline()){
+      await _service.removeFavoriteCaption(userId, captionId);
+    }
+
+    await loadAndSaveFavCaptions(userId);
+    notifyListeners();
   }
 
   bool isCaptionFavorite(String caption) {
-    return _favorites?.any((fav) => fav.caption == caption) ?? false;
+    return _favorites.any((fav) => fav.caption == caption);
   }
 
   Future<void> getAllFavCaptionsFromLocal(String userId, {bool isUpdate = true}) async {
     _favorites = await _dbHelper.getFavoriteCaptions(userId);
-    print('==========inside getAllFavCaptionsFromLocal _favorites length==========${_favorites?.length}');
+    print('==========inside getAllFavCaptionsFromLocal _favorites length==========${_favorites.length}');
     if(isUpdate){
       notifyListeners();
     }
